svg/text/parsing/text-decoration-color-computed.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml"><title>SVG CSS Text Decoration Test: getComputedStyle().textDecorationColor</title><metadata><h:link ref="https://drafts.csswg.org/css-text-decor-3/#text-decoration-color-property" rel="help"/><h:meta content="SVG text-decoration-color computed value is as specified." name="assert"/></metadata><style>#container{color:red;text-decoration-color:currentcolor}#target{color:#0f0}</style><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/computed-testcommon.js"/><script><![CDATA[

test_computed_value("text-decoration-color", "rgb(0, 0, 255)");
test_computed_value("text-decoration-color", "currentcolor", "rgb(0, 255, 0)");
test_computed_value("text-decoration-color", "inherit", "rgb(0, 255, 0)"); // currentcolor

  ]]></script></svg>
